requirements.txt
# AIOX Studio — Python Dependencies
# Install: pip install -r requirements.txt

# ── Core ──────────────────────────────────────
manim>=0.18.0
pyyaml>=6.0

# ── Phase 1: Static Generator ─────────────────
playwright>=1.44.0             # Screenshot pipeline (após instalar: playwright install chromium)

# ── Phase 2: Multi-Format Output ──────────────
Pillow>=10.3.0                 # Carousel, image composition
imageio>=2.34.0                # Multi-format animation support
opencv-python>=4.9.0           # Video frame processing
ffmpeg-python>=0.2.0           # FFmpeg Python wrapper
drawsvg>=2.3.0                 # SVG vector export

# ── Phase 3: Physics & Math ───────────────────
opensimplex>=0.4.5             # FBM noise — substitui harmônico em fields.py
numba>=0.59.0                  # JIT acceleration para ParticlePool (10k+ partículas)

# ── Phase 4: Shaders/GPU ──────────────────────
moderngl>=5.10.0               # GLSL shaders headless (ShaderLayer — fallback numpy se ausente)

# ── Phase 5: Skills & Extended Output ─────────
reportlab>=4.1.0               # PDF multi-página (pdf_generator.py)
